refactor: phase 2 — package restructuring
- create internal/paths/paths.go with XDG helpers (no zerolog dependency)
- create internal/config/config.go (moved from internal/core/config); add XDGConfigPath, ReadFileOrDefault; use paths.DataDir()/cache for backward compat
- delete internal/core/ entirely
- update imports in main.go, internal/gofind/app.go, internal/ui/styles.go

closes hc-1vmt46eb
